Add tests and sligh refactoring
